setMemo(), delMemo(), findMemo() 코드 추가

+++ app/src/main/java/com/example/semiproject_sample/bean/MemoBean.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class MemoBean {
 
-    public int memoID; //메모 고유 번호(메모를 찾기 위함)
+    public long memoID; //메모 고유 번호(메모를 찾기 위함)
     public String memoPicPath;
     public String memo;
     public String memoDate;
